# fix(editor): Improve `vue-tsc` command execution time by ~20 seconds (no-changelog)