--- a/packages/editor-ui/src/__tests__/data/index.ts
+++ b/packages/editor-ui/src/__tests__/data/index.ts
@@ -1,1 +1,2 @@
 export * from './canvas';
+export * from './nodeTypes';