--- a/packages/editor-ui/src/__tests__/defaults.ts
+++ b/packages/editor-ui/src/__tests__/defaults.ts
@@ -5,10 +5,7 @@ import {
 	CHAT_TRIGGER_NODE_TYPE,
 	MANUAL_TRIGGER_NODE_TYPE,
 } from '@/constants';
-import nodeTypesJson from '../../../nodes-base/dist/types/nodes.json';
-import aiNodeTypesJson from '../../../@n8n/nodes-langchain/dist/types/nodes.json';
-
-const allNodeTypes = [...nodeTypesJson, ...aiNodeTypesJson];
+import { allNodeTypes } from '@/__tests__/data';
 
 export function findNodeTypeDescriptionByName(name: string): INodeTypeDescription {
 	return allNodeTypes.find((node) => node.name === name) as INodeTypeDescription;