## Summary

Ran some performance benchmarks for `vue-tsc` and realized that the `__tests__/defaults.ts` file takes around 20 seconds to process. This happens because of the huge JSON files we're importing there. I've manually copied over the 4 node types that we need to reduce the time.

**Trace command:** `tsc --noEmit --generateTrace trace`
**Benchmark command:** `time pnpm run typecheck` (in editor-ui) 

<img width="1733" alt="image" src="https://github.com/n8n-io/n8n/assets/6179477/599f18be-d34a-4606-8ebc-f242a9a46652">

### Results:

**Before:**
```
pnpm run typecheck  59.30s user 5.63s system 147% cpu 44.005 total
```

**After:**
```
pnpm run typecheck  40.24s user 2.74s system 159% cpu 26.988 total
```

## Related Linear tickets, Github issues, and Community forum posts

N/A

## Review / Merge checklist

- [x] PR title and summary are descriptive. ([conventions](../blob/master/.github/pull_request_title_conventions.md)) <!--
   **Remember, the title automatically goes into the changelog.
   Use `(no-changelog)` otherwise.**
-->
- [ ] [Docs updated](https://github.com/n8n-io/n8n-docs) or follow-up ticket created.
- [ ] Tests included. <!--
   A bug is not considered fixed, unless a test is added to prevent it from happening again.
   A feature is not complete without tests.
-->
- [ ] PR Labeled with `release/backport` (if the PR is an urgent fix that needs to be backported)